requirements.txt
pydantic==1.10.11
newspaper3k==0.2.8
openai==0.27.8
fastapi==0.101.0
firebase-admin==6.2.0
pandas==2.0.3
httpx==0.24.1
transformers==4.31.0
google-search-results
python-dotenv==1.0.0
uvicorn==0.23.2
websockets==11.0.3
sentence-transformers==2.2.2
houndify==2.1.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-aiohttp==1.0.4
python-multipart
-r tools/requirements.txt